startup.sh: Ensure SCHEMAS is sorted
`find` doesn't sort the output, but if you don't sort it and two schema files are dependent on each other (schema 2 uses an attribute defined in schema 1) the startup of the container will fail. By sorting it you can know/rely on asciibetical sorting to ensure the schema files get included in the right order.